Develop

+++ src/main/java/br/com/exemplospring/exemplo/controller/UserController.java
@@ -4,6 +4,7 @@ import br.com.exemplospring.exemplo.entity.User;
 import br.com.exemplospring.exemplo.service.UserService;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import javafx.beans.property.adapter.JavaBeanBooleanPropertyBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,6 +13,9 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
+
+import static org.springframework.http.ResponseEntity.notFound;
+import static org.springframework.http.ResponseEntity.ok;
 
 @RestController
 @RequestMapping(value = "/users")
@@ -37,8 +41,7 @@ public class UserController {
         return userService.findAll();
     }
 
-    @GetMapping(value = "/{id}",
-            produces = "application/json")
+    @GetMapping(value = "/{id}", produces = "application/json")
     public ResponseEntity<User> GetById(@PathVariable Long id) {
         Optional<User> user = userService.findById(id);
         return user.map(value -> new ResponseEntity<>(value, HttpStatus.OK))
@@ -47,25 +50,24 @@ public class UserController {
 
     @PostMapping(produces = "application/json")
     public User Post(@RequestBody @Valid User user) {
-        return userService.save(user);
+       return userService.save(user);
     }
 
     @PutMapping(value = "/{id}")
-    public ResponseEntity<User> Put(@PathVariable @Valid Long id, @RequestBody User newUser) {
-        Optional<User> oldUser = userService.findById(id);
-        if (oldUser.isPresent()) {
-            User user = oldUser.get();
-            user.setName(newUser.getName());
-            userService.save(user);
-            return new ResponseEntity<>(user, HttpStatus.OK);
-        } else
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    public ResponseEntity<?> change(@PathVariable @Valid Long id, @RequestBody User user) {
+        if (userService.exist(id)) {
+            userService.change(id, user);
+            return ok(user);
+        }
+        return notFound().build();
     }
 
-    @DeleteMapping(value = "/{id}",
-            produces = "application/json")
-    public ResponseEntity<Object> delete(@PathVariable Long id) throws Exception {
-        userService.deleteById(id);
-        return ResponseEntity.noContent().build();
+    @DeleteMapping(value = "/{id}", produces = "application/json")
+    public ResponseEntity<User> delete(@PathVariable Long id) {
+        if (userService.exist(id)) {
+            userService.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
